algo

+++ src/algorithms/Dijkstra2.java
@@ -1,5 +1,5 @@
 package algorithms;
-// git Test
+
 import java.io.IOException;
 import java.util.Iterator;
 import java.util.LinkedList;
@@ -10,7 +10,6 @@ import org.graphstream.graph.Node;
 
 import application.graphicController;
 
-//Helge Timo Test push
 public class Dijkstra2 {
 	private Graph graph;
 	private Node source;
@@ -71,14 +70,14 @@ public class Dijkstra2 {
 	}
 
 	private Node nodeFromQ() {
-		int lowest = Integer.MAX_VALUE;
+		double lowest = Double.MAX_VALUE;
 		LinkedList<Node> queue2 = (LinkedList<Node>) queue.clone();
 		Node lowestNode = null;
 		int index = 0, counter = 0;
 
 		while (!queue2.isEmpty()) {
 			Node node = (Node) queue2.remove();
-			int i1 = (int) node.getNumber("distance");
+			double i1 = (double) node.getNumber("distance");
 			if (i1 < lowest) {
 				lowest = i1;
 				lowestNode = node;
@@ -107,10 +106,10 @@ public class Dijkstra2 {
 
 					 graphicController.colorEdge(edge, "red");
 					 graphicController.colorNode(node, "red");
-					int sEdgeDistance = (edge.getAttribute("weight"));
-					int distanceSource = (int) sourceNode.getNumber("distance");
-					int distanceNode = (int) node.getNumber("distance");
-					int edgeDistance = (sEdgeDistance);
+					double sEdgeDistance = (edge.getAttribute("weight"));
+					double distanceSource = sourceNode.getNumber("distance");
+					double distanceNode = node.getNumber("distance");
+					double edgeDistance = (sEdgeDistance);
 
 					if ((distanceSource + edgeDistance) < distanceNode) {
 						node.addAttribute("distance", (distanceSource + edgeDistance));
